refactor: relocate moveSync
