+++ src/common/path.ts
@@ -1,4 +1,4 @@
-import { cpSync, existsSync, lstatSync, mkdirSync, readdirSync, readlinkSync, rmSync, Stats } from "node:fs";
+import { existsSync, lstatSync, readdirSync, readlinkSync, Stats } from "node:fs";
 import { delimiter, dirname, join, resolve } from "node:path";
 import { fileURLToPath } from "node:url";
 
@@ -54,45 +54,6 @@ export function isSafeBaseName(value: string): boolean {
 	// Check the name does not contain an invalid character.
 	return !invalidCharacters.some((invalid) => value.includes(invalid));
 }
-
-/**
- * Synchronously moves the {@link source} to the {@link dest} path.
- * @param source Source path being moved.
- * @param dest Destination where the {@link source} will be moved to.
- * @param options Options that define the move.
- */
-export function moveSync(source: string, dest: string, options?: MoveOptions): void {
-	if (!existsSync(source)) {
-		throw new Error("Source does not exist");
-	}
-
-	if (!lstatSync(source).isDirectory()) {
-		throw new Error("Source must be a directory");
-	}
-
-	if (existsSync(dest)) {
-		if (options?.overwrite) {
-			rmSync(dest, { recursive: true });
-		} else {
-			throw new Error("Destination already exists");
-		}
-	}
-
-	// Ensure the new directory exists, copy the contents, and clean-up.
-	mkdirSync(dest, { recursive: true });
-	cpSync(source, dest, { recursive: true });
-	rmSync(source, { recursive: true });
-}
-
-/**
- * Defines how a path will be relocated.
- */
-type MoveOptions = {
-	/**
-	 * When the destination path already exists, it will be overwritten.
-	 */
-	overwrite?: boolean;
-};
 
 /**
  * Resolves the specified {@link path} relatives to the entry point.
